[geo] Took out GeoUrl parser from coding to ge0.
