+++ ge0/CMakeLists.txt
@@ -1,6 +1,8 @@
 project(ge0)
 
 set(SRC
+  geo_url_parser.cpp
+  geo_url_parser.hpp
   parser.cpp
   parser.hpp
   url_generator.cpp
