+++ ge0/ge0_tests/CMakeLists.txt
@@ -1,6 +1,7 @@
 project(ge0_tests)
 
 set(SRC
+  geo_url_tests.cpp
   parser_tests.cpp
   url_generator_tests.cpp
 )
