+++ indexer/editable_map_object.cpp
@@ -11,8 +11,6 @@
 #include "base/macros.hpp"
 #include "base/string_utils.hpp"
 
-#include "coding/url.hpp"
-
 #include <algorithm>
 #include <cctype>
 #include <cmath>
